# src/COP.API/wwwroot/swagger-original.json
{
  "swagger" : "2.0",
  "info" : {
    "version" : "0.1.0",
    "title" : "MONICA COP API"
  },
  "host" : "portal.monica-cloud.eu",
  "basePath" : "/cop/api",
  "schemes" : [ "https" ],
  "consumes" : [ "application/json" ],
  "produces" : [ "application/json" ],
  "paths" : {
    "/observation" : {
      "post" : {
        "tags" : [ "Observation" ],
        "description" : "add a new wearable",
        "parameters" : [ {
          "in" : "body",
          "name" : "thingid",
          "description" : "thing observation",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/Observation"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "schema" : {
              "$ref" : "#/definitions/GeneralResponse"
            }
          },
          "default" : {
            "description" : "Error",
            "schema" : {
              "$ref" : "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
    "/observations" : {
      "get" : {
        "tags" : [ "Observation" ],
        "description" : "list all observations",
        "parameters" : [ {
          "name" : "thingId",
          "in" : "query",
          "required" : false,
          "type" : "integer"
        }, {
          "name" : "datastreamID",
          "in" : "query",
          "required" : false,
          "type" : "boolean"
        } ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "schema" : {
              "$ref" : "#/definitions/GetObservationListResponse"
            }
          },
          "default" : {
            "description" : "Error",
            "schema" : {
              "$ref" : "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
    "/observations/{thingId}" : {
      "get" : {
        "tags" : [ "Observation" ],
        "description" : "get observation by id",
        "parameters" : [ {
          "name" : "thingId",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "schema" : {
              "$ref" : "#/definitions/Observation"
            }
          },
          "default" : {
            "description" : "Error",
            "schema" : {
              "$ref" : "#/definitions/ErrorResponse"
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "Observation" ],
        "description" : "delete observation by id",
        "parameters" : [ {
          "name" : "thingId",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "schema" : {
              "$ref" : "#/definitions/GeneralResponse"
            }
          },
          "default" : {
            "description" : "Error",
            "schema" : {
              "$ref" : "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
...
  },
  "definitions" : {
...
    "Observation" : {
      "type" : "object",
      "properties" : {
        "thingId" : {
          "type" : "integer",
          "description" : "Thing id"
        },
        "datastreamId" : {
          "type" : "string",
          "description" : "Datastream id"
        },
        "phenomenTime" : {
          "type" : "string",
          "format" : "date-time",
          "description" : "phenomen time"
        },
        "observationResult" : {
          "type" : "string",
          "description" : "The latest observation"
        }
      },
      "example" : {
        "datastreamId" : "datastreamId",
        "observationResult" : "observationResult",
        "phenomenTime" : "2000-01-23T04:56:07.000+00:00",
        "thingId" : 0
      }
    },
...
    "GetObservationListResponse" : {
      "type" : "array",
      "items" : {
        "$ref" : "#/definitions/Observation"
      }
    },
...
    "GeneralResponse" : {
      "type" : "object",
      "required" : [ "description", "success" ],
      "properties" : {
        "success" : {
          "type" : "boolean",
          "description" : "returns 1 if successful"
        },
        "description" : {
          "type" : "string",
          "description" : "a short comment"
        }
      },
      "example" : {
        "success" : true,
        "description" : "description"
      }
    },
    "ErrorResponse" : {
      "required" : [ "message" ],
      "properties" : {
        "message" : {
          "type" : "string"
        }
      }
    }
  }
}